Factor Linux specific functions out of zpool command
